requirements.txt
Django==2.2.12
django-autoslug==1.9.7
django-crispy-forms==1.9.0
djangorestframework==3.11.0
psycopg2==2.8.5
pytils==0.3
pytz==2020.1
rope==0.17.0
sqlparse==0.3.1
pre-commit==2.3.0
